Fixed a bug in compute_interference_graph() when using with multiple in-place operators.
Summary:
compute_interference_graph() was not able to handle the case when a blob is reused twice for operators supporting in-place parameters. For example, for the following network with operators Mul and Sub

(blob) -> [Mul] -> (blob) -> [Sub] -> (blob)

an incorrect edge will be added from [Sub] to [Mul] and causes nx.is_directed_acyclic_graph() to fail.

Reviewed By: ajtulloch

Differential Revision: D5271604

fbshipit-source-id: f6095b6f8e1dba556ba223a82c8170be7f744529