allow querying tensor device + tool to validate that all ops have tensors from correct devices (GPUs)
Summary:
Quite common, hard-to-debug, performance bug for multi-GPU training has been that operators have been passed tensors that reside on different GPU than what the op runs on. Since we have peer access enabled, this works, but is just much slower. With data parallel model this problem arises rarely as it has static analysis of the operators, but if someone bypassed DPM or uses FeedBlob with incorrect device options, this problem can happen.

To make debugging easier, I added device-field to tensor that stores the device information that allocated the memory. In addition, I added a function to go through operator inputs and outputs and compare their tensor device to the operator device. This check is run after first iteration with prof_dag only.

Also renamed ShapeCall to TensorInfoFun, as it now returns so much other info than the shape.

I think this is pretty safe diff, but do you find it problematic to add a new field to tensor?

Reviewed By: dzhulgakov

Differential Revision: D5335505

fbshipit-source-id: 511b6c122dff9a205f43951984868ffd40f7ac30