Sorted Order and Generalized Pattern Matching
Summary:
Pattern match currently only supports one type of pattern matching: connected components.

It will be useful to sometimes use different algorithms to pattern match, either a subset of the operators in order, or general non-connected subgraphs. While generalized pattern matching can match for all types, it is inefficient to use it when sorted order or connected component suffice.

You can can set the PatternMatchType to be one of the three options (it is connected by default), and Transform will use the associated algorithm.

We will need this for common subexpression elimination - specifically, sorted order matching.

Reviewed By: bwasti

Differential Revision: D5629321

fbshipit-source-id: 2104f2d4384fe4aba06a386881a08ca324f290a6