Fix stepworkspace sizing
Summary: When forward-only mode, we need only 2 workspaces. Errornously we sized the length of the workspace vector to 2 if it was different than 2. But if it was longer (because the step workspaces was shared by an non-forward-only op), we end up deleting the workspaces. With RNN Executor, this is a problem, because it held a reference to the deleted workspaces. Without RNN executor, we just ended recrearing the nets.

Reviewed By: jhcross

Differential Revision: D5654534

fbshipit-source-id: 1e6276e63453831747fee6a85c5057f01b89fde5