Update Caffe2 benchmark file to write text output
Summary:
The Caffe2 benchmarking framework can now compare the output of a model with some golden output. In order to do that, and reduce the dependency of the benchmarking framework and caffe2, the output is dumped as text format without any schema.

The output is read in by the benchmarking framework and perform the comparison.
Closes https://github.com/caffe2/caffe2/pull/1301

Reviewed By: bwasti

Differential Revision: D5992836

Pulled By: sf-wind

fbshipit-source-id: f6b403103949f4b9880c8372bbdc36966475a387